requirements.txt
flet==0.27.6
flet-cli==0.27.6
flet-core==0.24.1
flet-desktop==0.27.6
flet-geolocator==0.1.0
flet-runtime==0.24.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0